add navigation, page components and routing for a genre page with a filtered card container for each genre
