+++ src/components/moviedetails/MovieButton.jsx
@@ -1,6 +1,13 @@
 import styled from "styled-components"
 import { Link } from "react-router-dom"
 
+
+export const MovieButton = () => {
+  return (
+    <StyledLink to="/"><StyledArrow src="/arrow.png" alt="image of arrow" /><StyledText>Movies</StyledText></StyledLink>
+  );
+};
+
 const StyledArrow = styled.img`
   height: 22px;
   width: auto;
@@ -8,31 +15,23 @@ const StyledArrow = styled.img`
  @media (min-width: ${({ theme }) => theme.breakpoints.tablet}) {
     height: 22px;
   }
-`
+`;
+
 const StyledText = styled.div`
-margin: 3px;
-transition: .3s ease;
+  margin: 3px;
+  transition: .3s ease;
 
    &:hover {
     transform: translateX(5px)
-  }
- 
-   
-`
+  }  
+`;
 
 const StyledLink = styled(Link)`
   text-decoration: none;
   color: inherit;
   display: flex;
-   flex-direction: row;
-   gap: 3px;
+  flex-direction: row;
+  gap: 3px;
   margin: 25px 40px;
- 
-`
-
-export const MovieButton = () => {
-  return (
-    <StyledLink to="/"><StyledArrow src="/arrow.png" alt="image of arrow" /><StyledText>Movies</StyledText></StyledLink>
-  )
-}
+`;
